+++ frontend/src/components/MainToolbar.jsx
@@ -34,29 +34,32 @@ const ToolButton = styled.button`
 `;
 
 const Separator = styled.div`
-  width: 2px;
+  width: 20px;
   height: 20px;
-  border-left: 1px solid #808080;
-  border-right: 1px solid #fff;
+  /* border-left: 1px solid #808080; */
+  /* border-right: 1px solid #fff; */
   margin: 0 2px;
 `;
 
-const MainToolbar = ({ handleSessionLoad, handleStep, apiCall, toggleLogs, toggleDebugLogs, setActiveModal }) => {
+const MainToolbar = ({ handleSessionLoad, handleFileOpen, handleCloseTarget, handleStep, apiCall, openLogWindow, openDebugLogWindow, focusDisassembly, setActiveModal }) => {
 
   const handleRun = () => apiCall('/control/run', {}, 'POST');
   const handlePause = () => apiCall('/control/pause', {}, 'POST');
 
   return (
     <Toolbar>
-      <ToolButton onClick={handleSessionLoad} title="Reload Binary">↻</ToolButton>
+      <ToolButton onClick={() => setActiveModal('file_browser')} title="Open file">📂</ToolButton>
+      <ToolButton onClick={handleSessionLoad} title="Restart target">↻</ToolButton>
+      <ToolButton onClick={handleCloseTarget} title="Close target">X</ToolButton>
       <Separator />
       <ToolButton onClick={handleRun} title="Run (F9)">▶</ToolButton>
       <ToolButton onClick={handlePause} title="Pause (F12)">⏸</ToolButton>
       <ToolButton onClick={() => handleStep('step_into')} title="Step Into (F7)">⤵</ToolButton>
       <ToolButton onClick={() => handleStep('step_over')} title="Step Over (F8)">↷</ToolButton>
       <Separator />
-      <ToolButton onClick={toggleLogs} title="System Log (L)">L</ToolButton>
-      <ToolButton onClick={toggleDebugLogs} title="Debug Log (D)">D</ToolButton>
+      <ToolButton onClick={focusDisassembly} title="CPU Window (C)">C</ToolButton>
+      <ToolButton onClick={openDebugLogWindow} title="Log Window (L)">L</ToolButton>
+      <ToolButton onClick={openLogWindow} title="System Log (S)">S</ToolButton>
       <ToolButton onClick={() => setActiveModal('options')} title="Options (O)">O</ToolButton>
     </Toolbar>
   );